requirements.txt
Django==3.1.3
#mysqlclient==2.0.1
WeasyPrint==52.1
django-import-export==2.6.1
gunicorn==20.1.0
django-role-permissions==3.1.1
django-admin-autocomplete-filter==0.7.1
django-admin-list-filter-dropdown==1.0.3